Web/Backend/django-rest-fw/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
importlib_metadata==8.6.1
Markdown==3.8
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
zipp==3.21.0
